Split out spawners
Also made use of workspace dependencies
