+++ Cargo.toml
@@ -4,3 +4,27 @@ members = [
     "crates/*",
     "genesis_game"
 ]
+
+[workspace.dependencies]
+bevy = "0.9.1"
+bevy_egui = "0.17.1"
+bevy_rapier2d = {version = "0.19.0", features = ["debug-render"]}
+iyes_loopless = {version="0.9.1", features = ["fixedtimestep"]}
+derive_more = "0.99.17"
+derive-getters = "0.2.0"
+anyhow = "1.0.66"
+thiserror = "1.0.37"
+serde_derive = "1.0.147"
+serde = "1.0.149"
+serde_json = "1.0.89"
+rand = "0.8.5"
+rand_distr = "0.4.3"
+nalgebra = "0.31.4"
+ndarray = "0.15.6"
+num = "0.4.0"
+glam = "0.22.0"
+once_cell = "1.16.0"
+confy = "0.5.1"
+itertools = "0.10.5"
+bitvec = "1.0.1"
+cached = "0.40.0"
