+++ crates/genesis-brain/Cargo.toml
@@ -6,8 +6,8 @@ edition = "2021"
 license = "MIT"
 
 [dependencies]
-rand = "0.8.5"
-rand_distr = "0.4.3"
-thiserror = "1.0.37"
 genesis-util = {path = "../genesis-util/"}
-cached = "0.40.0"
+rand = { workspace = true }
+rand_distr = { workspace = true }
+thiserror = { workspace = true }
+cached = { workspace = true }
