+++ genesis_game/Cargo.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/BenGale93/genesis"
 readme = "../../README.md"
 
 [dependencies]
-bevy = "0.9.1"
-bevy_egui = "0.17.1"
-bevy_rapier2d = {version = "0.19.0", features = ["debug-render"]}
 genesis-lib = { path = "../crates/genesis-lib/"}
+bevy = {workspace= true}
+bevy_egui = {workspace= true}
+bevy_rapier2d = {workspace = true}
